veridica-data/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veridica-data"
version = "0.1.0"
description = "Extensible drug data pipeline for ChEMBL-centric pharmaceutical ML"
authors = [
    {name = "Veridica", email = "info@veridica.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]

dependencies = [
    # Core data processing
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "pyarrow>=10.0.0",
    
    # HTTP and API
    "requests>=2.28.0",
    "tenacity>=8.0.0",
    "python-dotenv>=0.19.0",
    
    # Chemical informatics
    "rdkit-pypi>=2022.9.1",
    "chembl-webresource-client>=0.10.0",
    
    # Data validation and schemas
    "pydantic>=1.10.0",
    
    # ML and analysis
    "scikit-learn>=1.1.0",
    
    # Progress and utilities
    "tqdm>=4.64.0",
    "orjson>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
]

[project.urls]
Homepage = "https://github.com/MahanAzadBeast/VerdicaGNOSISLEAP"
Repository = "https://github.com/MahanAzadBeast/VerdicaGNOSISLEAP"
Documentation = "https://github.com/MahanAzadBeast/VerdicaGNOSISLEAP/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true